refactor(cognitarium): move rdf stuff in its own crate
