+++ contracts/okp4-cognitarium/src/rdf/mod.rs
@@ -1,0 +1,5 @@
+mod serde;
+mod uri;
+
+pub use self::serde::*;
+pub use self::uri::*;
